requirements.txt
python-gitlab==2.10.1
pyhcl==0.4.4
flatdict==4.0.1
PyYAML==5.3.1
jmespath==0.10.0
kopf[full-auth]==1.35.3
pykube-ng==22.1.0
shortuuid==0.5.0
pytz==2021.3
python-dotenv==0.19.2
MarkupSafe==2.0.1
jinja2==3.0.3
redisgraph==2.4.4
requests==2.27.1
gitpython==3.1.27
